+++ bika/lims/content/analysisrequest.py
@@ -34,7 +34,6 @@ from bika.lims.config import PROJECTNAME
 from bika.lims.content.bikaschema import BikaSchema
 # Bika Interfaces
 from bika.lims.interfaces import IAnalysisRequest
-from bika.lims.interfaces import ISamplePrepWorkflow
 # Bika Permissions
 from bika.lims.permissions import EditARContact
 from bika.lims.permissions import ManageInvoices
@@ -114,7 +113,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'visible', 'edit': 'visible'},
                 'to_be_preserved': {'view': 'visible', 'edit': 'visible'},
                 'sample_due': {'view': 'visible', 'edit': 'visible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'visible', 'edit': 'visible'},
                 'attachment_due': {'view': 'visible', 'edit': 'visible'},
                 'to_be_verified': {'view': 'visible', 'edit': 'visible'},
@@ -163,7 +161,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'visible', 'edit': 'visible'},
                 'to_be_preserved': {'view': 'visible', 'edit': 'visible'},
                 'sample_due': {'view': 'visible', 'edit': 'visible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'visible', 'edit': 'visible'},
                 'attachment_due': {'view': 'visible', 'edit': 'visible'},
                 'to_be_verified': {'view': 'visible', 'edit': 'visible'},
@@ -287,7 +284,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'visible', 'edit': 'invisible'},
                 'to_be_preserved': {'view': 'visible', 'edit': 'invisible'},
                 'sample_due': {'view': 'visible', 'edit': 'invisible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'visible', 'edit': 'invisible'},
                 'attachment_due': {'view': 'visible', 'edit': 'invisible'},
                 'to_be_verified': {'view': 'visible', 'edit': 'invisible'},
@@ -327,7 +323,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'visible', 'edit': 'visible'},
                 'to_be_preserved': {'view': 'visible', 'edit': 'visible'},
                 'sample_due': {'view': 'visible', 'edit': 'visible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'visible', 'edit': 'visible'},
                 'attachment_due': {'view': 'visible', 'edit': 'visible'},
                 'to_be_verified': {'view': 'visible', 'edit': 'visible'},
@@ -367,7 +362,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'visible', 'edit': 'visible'},
                 'to_be_preserved': {'view': 'visible', 'edit': 'visible'},
                 'sample_due': {'view': 'visible', 'edit': 'visible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'visible', 'edit': 'visible'},
                 'attachment_due': {'view': 'visible', 'edit': 'visible'},
                 'to_be_verified': {'view': 'visible', 'edit': 'visible'},
@@ -405,7 +399,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'visible', 'edit': 'visible'},
                 'to_be_preserved': {'view': 'visible', 'edit': 'visible'},
                 'sample_due': {'view': 'visible', 'edit': 'visible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'visible', 'edit': 'visible'},
                 'attachment_due': {'view': 'visible', 'edit': 'visible'},
                 'to_be_verified': {'view': 'visible', 'edit': 'visible'},
@@ -457,7 +450,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'visible', 'edit': 'invisible'},
                 'to_be_preserved': {'view': 'visible', 'edit': 'invisible'},
                 'sample_due': {'view': 'visible', 'edit': 'invisible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'visible', 'edit': 'invisible'},
                 'attachment_due': {'view': 'visible', 'edit': 'invisible'},
                 'to_be_verified': {'view': 'visible', 'edit': 'invisible'},
@@ -520,7 +512,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'visible', 'edit': 'invisible'},
                 'to_be_preserved': {'view': 'visible', 'edit': 'invisible'},
                 'sample_due': {'view': 'visible', 'edit': 'invisible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'visible', 'edit': 'invisible'},
                 'attachment_due': {'view': 'visible', 'edit': 'invisible'},
                 'to_be_verified': {'view': 'visible', 'edit': 'invisible'},
@@ -560,7 +551,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'visible', 'edit': 'invisible'},
                 'to_be_preserved': {'view': 'visible', 'edit': 'invisible'},
                 'sample_due': {'view': 'visible', 'edit': 'invisible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'visible', 'edit': 'invisible'},
                 'attachment_due': {'view': 'visible', 'edit': 'invisible'},
                 'to_be_verified': {'view': 'visible', 'edit': 'invisible'},
@@ -597,7 +587,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'visible', 'edit': 'invisible'},
                 'to_be_preserved': {'view': 'visible', 'edit': 'invisible'},
                 'sample_due': {'view': 'visible', 'edit': 'invisible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'visible', 'edit': 'invisible'},
                 'attachment_due': {'view': 'visible', 'edit': 'invisible'},
                 'to_be_verified': {'view': 'visible', 'edit': 'invisible'},
@@ -673,7 +662,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'visible', 'edit': 'invisible'},
                 'to_be_preserved': {'view': 'visible', 'edit': 'invisible'},
                 'sample_due': {'view': 'visible', 'edit': 'invisible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'visible', 'edit': 'invisible'},
                 'attachment_due': {'view': 'visible', 'edit': 'invisible'},
                 'to_be_verified': {'view': 'visible', 'edit': 'invisible'},
@@ -712,7 +700,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'visible', 'edit': 'visible'},
                 'to_be_preserved': {'view': 'visible', 'edit': 'visible'},
                 'sample_due': {'view': 'visible', 'edit': 'visible'},
-                'sample_prep': {'view': 'visible', 'edit': 'visible'},
                 'sample_received': {'view': 'visible', 'edit': 'invisible'},
                 'attachment_due': {'view': 'visible', 'edit': 'invisible'},
                 'to_be_verified': {'view': 'visible', 'edit': 'invisible'},
@@ -781,7 +768,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'visible', 'edit': 'visible'},
                 'to_be_preserved': {'view': 'visible', 'edit': 'visible'},
                 'sample_due': {'view': 'visible', 'edit': 'visible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'visible', 'edit': 'visible'},
                 'attachment_due': {'view': 'visible', 'edit': 'visible'},
                 'to_be_verified': {'view': 'visible', 'edit': 'invisible'},
@@ -843,7 +829,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'invisible', 'edit': 'invisible'},
                 'to_be_preserved': {'view': 'invisible', 'edit': 'invisible'},
                 'sample_due': {'view': 'invisible', 'edit': 'invisible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'invisible', 'edit': 'invisible'},
                 'attachment_due': {'view': 'invisible', 'edit': 'invisible'},
                 'to_be_verified': {'view': 'invisible', 'edit': 'invisible'},
@@ -886,7 +871,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'visible', 'edit': 'visible'},
                 'to_be_preserved': {'view': 'visible', 'edit': 'visible'},
                 'sample_due': {'view': 'visible', 'edit': 'visible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'visible', 'edit': 'visible'},
                 'attachment_due': {'view': 'visible', 'edit': 'visible'},
                 'to_be_verified': {'view': 'visible', 'edit': 'visible'},
@@ -928,7 +912,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'visible', 'edit': 'visible'},
                 'to_be_preserved': {'view': 'visible', 'edit': 'visible'},
                 'sample_due': {'view': 'visible', 'edit': 'visible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'visible', 'edit': 'visible'},
                 'attachment_due': {'view': 'visible', 'edit': 'visible'},
                 'to_be_verified': {'view': 'visible', 'edit': 'visible'},
@@ -965,7 +948,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'visible', 'edit': 'visible'},
                 'to_be_preserved': {'view': 'visible', 'edit': 'visible'},
                 'sample_due': {'view': 'visible', 'edit': 'visible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'visible', 'edit': 'visible'},
                 'attachment_due': {'view': 'visible', 'edit': 'visible'},
                 'to_be_verified': {'view': 'visible', 'edit': 'visible'},
@@ -1002,7 +984,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'visible', 'edit': 'visible'},
                 'to_be_preserved': {'view': 'visible', 'edit': 'visible'},
                 'sample_due': {'view': 'visible', 'edit': 'visible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'visible', 'edit': 'visible'},
                 'attachment_due': {'view': 'visible', 'edit': 'visible'},
                 'to_be_verified': {'view': 'visible', 'edit': 'visible'},
@@ -1038,7 +1019,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'visible', 'edit': 'invisible'},
                 'to_be_preserved': {'view': 'visible', 'edit': 'invisible'},
                 'sample_due': {'view': 'visible', 'edit': 'invisible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'visible', 'edit': 'invisible'},
                 'attachment_due': {'view': 'visible', 'edit': 'invisible'},
                 'to_be_verified': {'view': 'visible', 'edit': 'invisible'},
@@ -1079,7 +1059,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'visible', 'edit': 'visible'},
                 'to_be_preserved': {'view': 'visible', 'edit': 'visible'},
                 'sample_due': {'view': 'visible', 'edit': 'visible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'visible', 'edit': 'invisible'},
                 'attachment_due': {'view': 'visible', 'edit': 'invisible'},
                 'to_be_verified': {'view': 'visible', 'edit': 'invisible'},
@@ -1122,7 +1101,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'visible', 'edit': 'visible'},
                 'to_be_preserved': {'view': 'visible', 'edit': 'visible'},
                 'sample_due': {'view': 'visible', 'edit': 'visible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'visible', 'edit': 'invisible'},
                 'attachment_due': {'view': 'visible', 'edit': 'invisible'},
                 'to_be_verified': {'view': 'visible', 'edit': 'invisible'},
@@ -1159,7 +1137,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'visible', 'edit': 'visible'},
                 'to_be_preserved': {'view': 'visible', 'edit': 'visible'},
                 'sample_due': {'view': 'visible', 'edit': 'visible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'visible', 'edit': 'visible'},
                 'attachment_due': {'view': 'visible', 'edit': 'visible'},
                 'to_be_verified': {'view': 'visible', 'edit': 'visible'},
@@ -1230,7 +1207,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'visible', 'edit': 'invisible'},
                 'to_be_preserved': {'view': 'visible', 'edit': 'invisible'},
                 'sample_due': {'view': 'visible', 'edit': 'invisible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'visible', 'edit': 'invisible'},
                 'attachment_due': {'view': 'visible', 'edit': 'invisible'},
                 'to_be_verified': {'view': 'visible', 'edit': 'invisible'},
@@ -1271,7 +1247,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'visible', 'edit': 'invisible'},
                 'to_be_preserved': {'view': 'visible', 'edit': 'invisible'},
                 'sample_due': {'view': 'visible', 'edit': 'invisible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'visible', 'edit': 'invisible'},
                 'attachment_due': {'view': 'visible', 'edit': 'invisible'},
                 'to_be_verified': {'view': 'visible', 'edit': 'invisible'},
@@ -1307,7 +1282,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'visible', 'edit': 'visible'},
                 'to_be_preserved': {'view': 'visible', 'edit': 'visible'},
                 'sample_due': {'view': 'visible', 'edit': 'visible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'visible', 'edit': 'visible'},
                 'attachment_due': {'view': 'visible', 'edit': 'visible'},
                 'to_be_verified': {'view': 'visible', 'edit': 'visible'},
@@ -1341,7 +1315,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'visible', 'edit': 'visible'},
                 'to_be_preserved': {'view': 'visible', 'edit': 'visible'},
                 'sample_due': {'view': 'visible', 'edit': 'visible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'visible', 'edit': 'invisible'},
                 'attachment_due': {'view': 'visible', 'edit': 'invisible'},
                 'to_be_verified': {'view': 'visible', 'edit': 'invisible'},
@@ -1447,7 +1420,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'invisible', 'edit': 'invisible'},
                 'to_be_preserved': {'view': 'invisible', 'edit': 'invisible'},
                 'sample_due': {'view': 'invisible', 'edit': 'invisible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'visible', 'edit': 'invisible'},
                 'attachment_due': {'view': 'visible', 'edit': 'invisible'},
                 'to_be_verified': {'view': 'visible', 'edit': 'invisible'},
@@ -1479,7 +1451,6 @@ schema = BikaSchema.copy() + Schema((
                 'sampled': {'view': 'invisible', 'edit': 'invisible'},
                 'to_be_preserved': {'view': 'invisible', 'edit': 'invisible'},
                 'sample_due': {'view': 'invisible', 'edit': 'invisible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
                 'sample_received': {'view': 'invisible', 'edit': 'invisible'},
                 'attachment_due': {'view': 'invisible', 'edit': 'invisible'},
                 'to_be_verified': {'view': 'invisible', 'edit': 'invisible'},
@@ -1754,41 +1725,6 @@ schema = BikaSchema.copy() + Schema((
         ),
     ),
 
-    StringField(
-        'PreparationWorkflow',
-        mode="rw",
-        read_permission=View,
-        write_permission=ModifyPortalContent,
-        vocabulary='getPreparationWorkflows',
-        acquire=True,
-        widget=BikaSelectionWidget(
-            format="select",
-            label=_("Preparation Workflow"),
-            visible={
-                'edit': 'visible',
-                'view': 'visible',
-                'add': 'edit',
-                'header_table': 'visible',
-                'sample_registered': {
-                    'view': 'visible', 'edit': 'visible', 'add': 'edit'},
-                'to_be_sampled': {'view': 'visible', 'edit': 'visible'},
-                'sampled': {'view': 'visible', 'edit': 'visible'},
-                'to_be_preserved': {'view': 'visible', 'edit': 'visible'},
-                'sample_due': {'view': 'visible', 'edit': 'visible'},
-                'sample_prep': {'view': 'visible', 'edit': 'invisible'},
-                'sample_received': {'view': 'visible', 'edit': 'invisible'},
-                'attachment_due': {'view': 'visible', 'edit': 'invisible'},
-                'to_be_verified': {'view': 'visible', 'edit': 'invisible'},
-                'verified': {'view': 'visible', 'edit': 'invisible'},
-                'published': {'view': 'visible', 'edit': 'invisible'},
-                'invalid': {'view': 'visible', 'edit': 'invisible'},
-                'rejected': {'view': 'visible', 'edit': 'invisible'},
-            },
-            render_own_label=True,
-        ),
-    ),
-
-
     # For comments or results interpretation
     # Old one, to be removed because of the incorporation of
     # ResultsInterpretationDepts (due to LIMS-1628)
@@ -1867,7 +1803,7 @@ schema.moveField('ResultsInterpretationDepts', pos='bottom')
 
 
 class AnalysisRequest(BaseFolder):
-    implements(IAnalysisRequest, ISamplePrepWorkflow)
+    implements(IAnalysisRequest)
     security = ClassSecurityInfo()
     displayContentsTab = False
     schema = schema
@@ -2666,19 +2602,6 @@ class AnalysisRequest(BaseFolder):
     def getSamplers(self):
         return getUsers(self, ['Sampler', ])
 
-    def getPreparationWorkflows(self):
-        """Return a list of sample preparation workflows.  These are identified
-        by scanning all workflow IDs for those beginning with "sampleprep".
-        """
-        wf = self.portal_workflow
-        ids = wf.getWorkflowIds()
-        sampleprep_ids = [wid for wid in ids if wid.startswith('sampleprep')]
-        prep_workflows = [['', ''], ]
-        for workflow_id in sampleprep_ids:
-            workflow = wf.getWorkflowById(workflow_id)
-            prep_workflows.append([workflow_id, workflow.title])
-        return DisplayList(prep_workflows)
-
     def getDepartments(self):
         """Returns a list of the departments assigned to the Analyses
         from this Analysis Request
@@ -2996,14 +2919,6 @@ class AnalysisRequest(BaseFolder):
     @security.public
     def guard_receive_transition(self):
         return guards.receive(self)
-
-    @security.public
-    def guard_sample_prep_transition(self):
-        return guards.sample_prep(self)
-
-    @security.public
-    def guard_sample_prep_complete_transition(self):
-        return guards.sample_prep_complete(self)
 
     @security.public
     def guard_schedule_sampling_transition(self):